Move test definitions to separate module
- All test-related structs and traits moved to mod file

- Tests for each operation moved to their respective files
